refactor(doctor): 拆分体检契约与运行时声明
