+++ src/doctor/report-schema.ts
@@ -1,0 +1,3 @@
+/** Bumped whenever DoctorReport schema changes in a way CI consumers should
+ *  be able to detect. CI can pin/check this when parsing the JSON. */
+export const DOCTOR_REPORT_SCHEMA_VERSION = '3.0.0';
