+++ src/doctor/rule-kind.ts
@@ -1,0 +1,5 @@
+import type { ComposerRule, DoctorRuleLike } from './contracts.js';
+
+export function isComposerRule(rule: DoctorRuleLike): rule is ComposerRule {
+  return (rule as ComposerRule).ruleKind === 'composer';
+}
