+++ src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './shared.js';
-export * from './doctor.js';
 export * from './skill-index.js';
