+++ test/artifact-graph/doctor-graph.test.ts
@@ -10,7 +10,7 @@ import {
   removeDoctorGraphSidecars,
   renderDoctorEvidenceCard,
 } from '../../src/artifact-graph/doctor.js';
-import type { DoctorReport } from '../../src/types/index.js';
+import type { DoctorReport } from '../../src/doctor/contracts.js';
 
 function makeReport(cwd: string, skillPath: string): DoctorReport {
   return {
