+++ test/doctor/health/composer.test.ts
@@ -12,7 +12,7 @@ import {
 import type { Artifact } from '../../../src/artifacts/contracts.js';
 import type { ExecutorFn } from '../../../src/executors/contracts/ports.js';
 import type { ExecResult } from '../../../src/executors/contracts/result.js';
-import type { DoctorContext } from '../../../src/types/doctor.js';
+import type { DoctorContext } from '../../../src/doctor/contracts.js';
 import type { HealthDimensionSpec } from '../../../src/doctor/health/dimension-spec.js';
 
 const stubDim = (id: string, severity: 'fatal' | 'warn' | 'info' = 'warn'): HealthDimensionSpec => ({
