+++ test/scripts/legacy-evaluation-removal.test.ts
@@ -9,6 +9,7 @@ const FORBIDDEN_FILES = [
   'src/types/eval.ts',
   'src/types/judge.ts',
   'src/types/diagnosis.ts',
+  'src/types/doctor.ts',
   'src/eval-workflows/run-evaluation.ts',
   'src/eval-workflows/evaluation-pipeline.ts',
   'src/eval-workflows/batch-evaluation-workflow.ts',
@@ -27,6 +28,7 @@ const FORBIDDEN_SPECIFIER_FRAGMENTS = [
   '/types/eval',
   '/types/judge',
   '/types/diagnosis',
+  '/types/doctor',
   '/eval-workflows/run-evaluation',
   '/eval-workflows/evaluation-pipeline',
   '/server/report-store',
