+++ test/server/cross-project-doctor-observe.test.ts
@@ -15,7 +15,7 @@ import {
 } from '../../src/measurement-artifacts/discovery-index.js';
 import { reportFileName } from '../../src/measurement-artifacts/file-names.js';
 import { pruneDoctorHistory } from '../../src/cli/commands/doctor.js';
-import type { DoctorReport } from '../../src/types/index.js';
+import type { DoctorReport } from '../../src/doctor/contracts.js';
 
 type DoctorCardInput = Parameters<typeof writeDoctorIndex>[0];
 type ObserveSource = Parameters<typeof writeObserveIndex>[0];
